Compute V2: migrate rescue to separate package
Move the Rescue action of the Compute service to its own package to
leverage common Gophercloud structuring.
