+++ openstack/compute/v2/extensions/rescueunrescue/doc.go
@@ -1,0 +1,1 @@
+package rescueunrescue
