+++ openstack/compute/v2/extensions/rescueunrescue/testing/doc.go
@@ -1,0 +1,1 @@
+package testing
