+++ openstack/compute/v2/extensions/rescueunrescue/urls.go
@@ -1,0 +1,7 @@
+package rescueunrescue
+
+import "github.com/gophercloud/gophercloud"
+
+func actionURL(client *gophercloud.ServiceClient, id string) string {
+	return client.ServiceURL("servers", id, "action")
+}
