+++ openstack/compute/v2/servers/testing/requests_test.go
@@ -383,20 +383,6 @@ func TestRevertResize(t *testing.T) {
 	th.AssertNoErr(t, res.Err)
 }
 
-func TestRescue(t *testing.T) {
-	th.SetupHTTP()
-	defer th.TeardownHTTP()
-
-	HandleServerRescueSuccessfully(t)
-
-	res := servers.Rescue(client.ServiceClient(), "1234asdf", servers.RescueOpts{
-		AdminPass: "1234567890",
-	})
-	th.AssertNoErr(t, res.Err)
-	adminPass, _ := res.Extract()
-	th.AssertEquals(t, "1234567890", adminPass)
-}
-
 func TestGetMetadatum(t *testing.T) {
 	th.SetupHTTP()
 	defer th.TeardownHTTP()
